refactor: clean up declarations and javadocs per IntelliJ - Make AutoPaths final - Fix IntakeSequence javadocs - Rename RevBlinken to RevBlinkin (actual product name)
