+++ src/main/java/org/team1540/robot2022/utils/AutoHelper.java
@@ -154,17 +154,18 @@ public class AutoHelper {
     }
 
     public static class AutoPath {
-        public static AutoPath auto1Ball = new AutoPath("2ball.posA.path1.wpilib.json", 2200, 2500, true);
-        public static AutoPath auto2Ball1A = new AutoPath("2ball.posA.path1.wpilib.json", 2200, 2500, true);
-        public static AutoPath auto2Ball1B = new AutoPath("2ball.posB.path1.wpilib.json", 2200, 2500, true);
-        public static AutoPath auto3Ball2 = new AutoPath("3ball.posA.path2.wpilib.json", 2500, 2800, true);
-        public static AutoPath auto4Ball2 = new AutoPath("4ball.posA.path2.wpilib.json", 2400, 2800, true);
-        public static AutoPath auto5Ball2 = new AutoPath("5ball.posA.path2.wpilib.json", 2600, 2900, true);
-        public static AutoPath auto5Ball3 = new AutoPath("5ball.posA.path3.wpilib.json", 2400, 2800, true);
-        public Trajectory trajectory;
-        public boolean hoodState;
-        public double frontSetpoint, rearSetpoint;
-        public String name;
+        public static final AutoPath auto1Ball = new AutoPath("2ball.posA.path1.wpilib.json", 2200, 2500, true);
+        public static final AutoPath auto2Ball1A = new AutoPath("2ball.posA.path1.wpilib.json", 2200, 2500, true);
+        public static final AutoPath auto2Ball1B = new AutoPath("2ball.posB.path1.wpilib.json", 2200, 2500, true);
+        public static final AutoPath auto3Ball2 = new AutoPath("3ball.posA.path2.wpilib.json", 2500, 2800, true);
+        public static final AutoPath auto4Ball2 = new AutoPath("4ball.posA.path2.wpilib.json", 2400, 2800, true);
+        public static final AutoPath auto5Ball2 = new AutoPath("5ball.posA.path2.wpilib.json", 2600, 2900, true);
+        public static final AutoPath auto5Ball3 = new AutoPath("5ball.posA.path3.wpilib.json", 2400, 2800, true);
+
+        public final Trajectory trajectory;
+        public final boolean hoodState;
+        public final double frontSetpoint, rearSetpoint;
+        public final String name;
 
         private AutoPath(String pathName, double frontSetpoint, double rearSetpoint, boolean hoodState) {
             this.trajectory = AutoHelper.getTrajectory(pathName);
